django_3/requirements.txt
Django==5.0
djangorestframework==3.14.0
django-cors-headers==4.3.1
coreapi==2.3.3
PyYAML==6.0.1
uritemplate==4.1.1
markdown==3.5.1
django-filter==23.5
